requirements.txt
aioredis==1.3.1
anyascii==0.3.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
contractions==0.1.66
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.2
django-allauth==0.48.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
joblib==1.1.0
learn==1.0.0
msgpack==1.0.3
nltk==3.7
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyahocorasick==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.1.4
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
textsearch==0.0.21
threadpoolctl==3.1.0
tqdm==4.62.3
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
zope.interface==5.4.0
